refactor: fold core types into concepts, delete deployment-config crate
With the manifest config model gone, deployment-config held only the core
id/naming/env-var types that concepts already re-exported verbatim. Move
component_id, naming, env_var and their rusqlite/postgres trait impls into
concepts (concepts::component_id / ::naming become real modules, so downstream
import paths are unchanged) and delete the crate.

wasm-workers no longer depends on the crate; the binary's env-var re-export now
points at concepts.
