+++ Cargo.toml
@@ -12,7 +12,6 @@ version.workspace = true
 
 [dependencies]
 concepts.workspace = true
-deployment-config.workspace = true
 db-postgres.workspace = true
 db-sqlite.workspace = true
 executor.workspace = true
@@ -152,7 +151,6 @@ members = [
     "crates/bench",
     "crates/concepts",
     "crates/db-common",
-    "crates/deployment-config",
     "crates/db-postgres",
     "crates/db-sqlite",
     "crates/embedded-assets",
@@ -214,7 +212,6 @@ concepts = { package = "obeli-sk-concepts", path = "crates/concepts", version = 
 db-common = { package = "obeli-sk-db-common", path = "crates/db-common", version = "0.41.2" }
 db-postgres = { package = "obeli-sk-db-postgres", path = "crates/db-postgres", version = "0.41.2" }
 db-sqlite = { package = "obeli-sk-db-sqlite", path = "crates/db-sqlite", version = "0.41.2" }
-deployment-config = { package = "obeli-sk-deployment-config", path = "crates/deployment-config", version = "0.41.2" }
 db-tests = { package = "obeli-db-tests", path = "crates/testing/db-tests" }
 embedded-assets = { package = "obeli-sk-embedded-assets", path = "crates/embedded-assets", version = "0.41.2" }
 executor = { package = "obeli-sk-executor", path = "crates/executor", version = "0.41.2" }
