+++ crates/concepts/src/component_id.rs
