--- crates/deployment-config/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "obeli-sk-deployment-config"
-description = "Deployment configuration schema, transient resolution types, and core naming types of Obelisk."
-
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-arbitrary = { workspace = true, optional = true }
-const_format.workspace = true
-derive_more.workspace = true
-hashbrown.workspace = true
-oci-client.workspace = true
-reqwest.workspace = true
-rusqlite = { workspace = true, optional = true }
-schemars.workspace = true
-serde.workspace = true
-serde_with.workspace = true
-strum.workspace = true
-thiserror.workspace = true
-tokio-postgres = { workspace = true, optional = true }
-tracing.workspace = true
-wit-parser.workspace = true
-
-[features]
-test = ["dep:arbitrary"]
-rusqlite = ["dep:rusqlite"]
-postgres = ["dep:tokio-postgres"]
-
-[lints]
-workspace = true
